Renamed/moved the scheduler and workpool files
